+++ RELEASE_NOTES.rst
@@ -19,6 +19,7 @@ improvements, and is now subject to static type checking with |mypy|.
   estimators are compatible with the wrapper class
 - FIX: base :class:`.LGBMClassifierDF` and :class:`.XGBClassifierDF` on the
   the correct wrapper class :class:`.ClassifierWrapperDF`
+- FIX: support array-like values for parameter ``drop`` of :class:`.OneHotEncoderDF`
 
 
 2.0.0